combine dummy consensus and round robin
